Deduplicate redundancy on migrations
